chore(api,worker): dedupe _crypto.py into shared checks.crypto module (#83)
